Host&Port configurable by env var and config
