--- demo/src/main/resources/application.conf
@@ -1,10 +1,0 @@
-rpc {
-
-  client {
-    host = "localhost"
-    host = ${?RPC_HOST}
-    port = 50051
-    port = ${?RPC_PORT}
-  }
-
-}
