+++ frees-rpc/src/main/resources/application.conf
@@ -1,8 +1,8 @@
 rpc {
 
-  server {
-    port = 50051
-    port = ${?RPC_PORT}
-  }
+  host = "localhost"
+  host = ${?RPC_HOST}
+  port = 8080
+  port = ${?RPC_PORT}
 
 }
