+++ frees-rpc/src/main/scala/client/package.scala
@@ -1,7 +1,9 @@
 package metrifier
-package benchmark
+package rpc
 
+import cats.implicits._
 import freestyle.implicits._
+import freestyle.config.implicits._
 import freestyle.asyncCatsEffect.implicits._
 import freestyle.rpc.client._
 import freestyle.rpc.client.implicits._
@@ -11,12 +13,20 @@ import metrifier.rpc._
 import metrifier.rpc.protocols._
 
 import scala.concurrent.Future
+import scala.util.{Failure, Success, Try}
 
-package object rpc {
+package object client {
 
   trait ClientConf {
 
-    val channelFor: ManagedChannelFor = ManagedChannelForAddress("localhost", 50051)
+    val channelFor: ManagedChannelFor =
+      ConfigForAddress[ChannelConfig.Op]("rpc.host", "rpc.port")
+        .interpret[Try] match {
+        case Success(c) => c
+        case Failure(e) =>
+          e.printStackTrace()
+          throw new RuntimeException("Unable to load the client configuration", e)
+      }
 
     val channelConfigList: List[ManagedChannelConfig] = List(UsePlaintext(true))
