+++ http/src/main/resources/application.conf
@@ -1,0 +1,8 @@
+http {
+
+  host = "localhost"
+  host = ${?HTTP_HOST}
+  port = 8080
+  port = ${?HTTP_PORT}
+
+}
