--- http/src/main/scala/server/http.scala
@@ -1,9 +1,0 @@
-package metrifier
-package http
-package server
-
-object HttpConf {
-
-  val host = "0.0.0.0"
-  val port = 8080
-}
